requirements.txt
accelerate==0.23.0
albumentations==1.3.1
colorama==0.4.6
einops==0.8.0
matplotlib==3.8.0
numpy==1.26.4
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.1.1
Pillow==10.0.1
PyYAML==6.0
thop==0.1.1.post2209072238
tqdm==4.66.1
tensorboard==2.14.1